ISSUE #2186: Consolidated CustomGear and ImportedGear to Gear
