+++ src/Application/Import/StravaImport/ImportGear/GearImportStatus.php
@@ -5,19 +5,19 @@ declare(strict_types=1);
 namespace App\Application\Import\StravaImport\ImportGear;
 
 use App\Domain\Gear\Gear;
-use App\Domain\Gear\ImportedGear\ImportedGearRepository;
+use App\Domain\Gear\GearRepository;
 
 final readonly class GearImportStatus
 {
     public function __construct(
-        private ImportedGearRepository $importedGearRepository,
+        private GearRepository $gearRepository,
     ) {
     }
 
     public function isComplete(): bool
     {
-        $stravaGearIdsOnActivities = $this->importedGearRepository->findUniqueStravaGearIds(null);
-        $importedGears = $this->importedGearRepository->findAll();
+        $stravaGearIdsOnActivities = $this->gearRepository->findUniqueStravaGearIds(null);
+        $importedGears = $this->gearRepository->findAll();
 
         foreach ($stravaGearIdsOnActivities as $gearId) {
             if (!$importedGears->getByGearId($gearId) instanceof Gear) {
