--- src/Domain/Gear/CustomGear/CustomGearRepository.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Gear\CustomGear;
-
-use App\Domain\Gear\Gears;
-
-interface CustomGearRepository
-{
-    public function save(CustomGear $gear): void;
-
-    public function findAll(): Gears;
-
-    public function findAllUsed(): Gears;
-
-    public function removeAll(): void;
-}
