--- src/Domain/Gear/ImportedGear/ImportedGearRepository.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Gear\ImportedGear;
-
-use App\Domain\Activity\ActivityIds;
-use App\Domain\Gear\GearId;
-use App\Domain\Gear\GearIds;
-use App\Domain\Gear\Gears;
-
-interface ImportedGearRepository
-{
-    public function save(ImportedGear $gear): void;
-
-    public function findAll(): Gears;
-
-    public function findAllUsed(): Gears;
-
-    public function find(GearId $gearId): ImportedGear;
-
-    public function findUniqueStravaGearIds(?ActivityIds $restrictToActivityIds): GearIds;
-}
