--- tests/Domain/Gear/CustomGear/DbalCustomGearRepositoryTest.php
@@ -1,165 +1,0 @@
-<?php
-
-namespace App\Tests\Domain\Gear\CustomGear;
-
-use App\Domain\Activity\ActivityId;
-use App\Domain\Activity\ActivityRepository;
-use App\Domain\Activity\ActivityType;
-use App\Domain\Activity\ActivityTypes;
-use App\Domain\Activity\ActivityWithRawData;
-use App\Domain\Gear\CustomGear\CustomGearRepository;
-use App\Domain\Gear\CustomGear\DbalCustomGearRepository;
-use App\Domain\Gear\GearId;
-use App\Domain\Gear\Gears;
-use App\Domain\Gear\ImportedGear\ImportedGearRepository;
-use App\Infrastructure\ValueObject\Measurement\Length\Kilometer;
-use App\Infrastructure\ValueObject\Measurement\Length\Meter;
-use App\Infrastructure\ValueObject\Measurement\Time\Seconds;
-use App\Tests\ContainerTestCase;
-use App\Tests\Domain\Activity\ActivityBuilder;
-use App\Tests\Domain\Gear\ImportedGear\ImportedGearBuilder;
-use Spatie\Snapshots\MatchesSnapshots;
-
-class DbalCustomGearRepositoryTest extends ContainerTestCase
-{
-    use MatchesSnapshots;
-
-    private CustomGearRepository $customGearRepository;
-
-    public function testSave(): void
-    {
-        $gear = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(1))
-            ->withDistanceInMeter(Meter::from(1230))
-            ->build();
-        $this->customGearRepository->save($gear);
-
-        $this->assertMatchesJsonSnapshot(
-            $this->getConnection()->executeQuery('SELECT * FROM GEAR')->fetchAllAssociative()
-        );
-    }
-
-    public function testFindAll(): void
-    {
-        $activityRepository = $this->getContainer()->get(ActivityRepository::class);
-
-        $gearOne = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(1))
-            ->withDistanceInMeter(Meter::from(1230))
-            ->build();
-        $this->customGearRepository->save($gearOne);
-        $gearTwo = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(2))
-            ->withDistanceInMeter(Meter::from(10230))
-            ->build();
-        $this->customGearRepository->save($gearTwo);
-        $gearThree = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(3))
-            ->withDistanceInMeter(Meter::from(230))
-            ->build();
-        $this->customGearRepository->save($gearThree);
-        $gearFour = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(4))
-            ->withDistanceInMeter(Meter::from(100230))
-            ->withIsRetired(true)
-            ->build();
-        $this->customGearRepository->save($gearFour);
-
-        $activityRepository->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('1'))
-                ->withGearId(GearId::fromUnprefixed(2))
-                ->withDistance(Kilometer::from(15))
-                ->withElevation(Meter::from(150))
-                ->withMovingTimeInSeconds(5400)
-                ->build(),
-            []
-        ));
-        $activityRepository->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('2'))
-                ->withGearId(GearId::fromUnprefixed(2))
-                ->withDistance(Kilometer::from(25))
-                ->withElevation(Meter::from(250))
-                ->withMovingTimeInSeconds(9000)
-                ->build(),
-            []
-        ));
-
-        $this->assertEquals(
-            Gears::fromArray([
-                CustomGearBuilder::fromDefaults()
-                    ->withGearId(GearId::fromUnprefixed(2))
-                    ->withDistanceInMeter(Meter::from(40000))
-                    ->withMovingTime(Seconds::from(14400))
-                    ->withElevation(Meter::from(400))
-                    ->withNumberOfActivities(2)
-                    ->build()
-                    ->withActivityTypes(ActivityTypes::fromArray([ActivityType::RIDE])),
-                CustomGearBuilder::fromDefaults()
-                    ->withGearId(GearId::fromUnprefixed(1))
-                    ->withDistanceInMeter(Meter::zero())
-                    ->build(),
-                CustomGearBuilder::fromDefaults()
-                    ->withGearId(GearId::fromUnprefixed(3))
-                    ->withDistanceInMeter(Meter::zero())
-                    ->build(),
-                CustomGearBuilder::fromDefaults()
-                    ->withGearId(GearId::fromUnprefixed(4))
-                    ->withDistanceInMeter(Meter::zero())
-                    ->withIsRetired(true)
-                    ->build(),
-            ]),
-            $this->customGearRepository->findAll()
-        );
-    }
-
-    public function testUpdate(): void
-    {
-        $gear = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(1))
-            ->build();
-        $this->customGearRepository->save($gear);
-
-        $this->customGearRepository->save($gear->withName('Updated gear')->withIsRetired(true));
-
-        $this->assertMatchesJsonSnapshot(
-            $this->getConnection()->executeQuery('SELECT * FROM GEAR')->fetchAllAssociative()
-        );
-    }
-
-    public function testRemoveAll(): void
-    {
-        $gearOne = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(1))
-            ->withDistanceInMeter(Meter::from(1230))
-            ->build();
-        $this->customGearRepository->save($gearOne);
-        $gearTwo = CustomGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(2))
-            ->withDistanceInMeter(Meter::from(10230))
-            ->build();
-        $this->customGearRepository->save($gearTwo);
-        $gearThree = ImportedGearBuilder::fromDefaults()
-            ->withGearId(GearId::fromUnprefixed(3))
-            ->withDistanceInMeter(Meter::from(230))
-            ->build();
-        $this->getContainer()->get(ImportedGearRepository::class)->save($gearThree);
-
-        $this->customGearRepository->removeAll();
-
-        $this->assertMatchesJsonSnapshot(
-            $this->getConnection()->executeQuery('SELECT * FROM GEAR')->fetchAllAssociative()
-        );
-    }
-
-    #[\Override]
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->customGearRepository = new DbalCustomGearRepository(
-            $this->getConnection()
-        );
-    }
-}
